Trait pallet_accounts::pallet::Config

pub trait Config: Config + CreateSignedTransaction<Call<Self>> + Config + Config + Config + Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
    type EthChainId: Get<u64>;
    type EthConfirmations: Get<u64>;
    type EthFragContract: EthFragContract;
    type Threshold: Get<u64>;
    type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
    type TicketsAssetId: Get<<Self as Config>::AssetId>;
    type InitialPercentageTickets: Get<u8>;
    type InitialPercentageNova: Get<u8>;
    type USDEquivalentAmount: Get<u128>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type WeightInfo: WeightInfo

Weight functions needed for pallet_accounts.

type EthChainId: Get<u64>

The Ethereum Chain ID that the Fragnova-owned Ethereum Smart Contract is deployed on. This should be the Ethereum Mainnet’s Chain ID.

type EthConfirmations: Get<u64>

The number of confirmations required to consider a transaction on the Ethereum Blockchain final (https://www.youtube.com/watch?v=gP5zcHD8tJU)

type EthFragContract: EthFragContract

Traits of the FRAG Token Smart Contract on the Ethereum Blockchain

type Threshold: Get<u64>

Number of votes needed to do something (问Gio)

type AuthorityId: AppCrypto<Self::Public, Self::Signature>

The identifier type for an offchain worker.

type TicketsAssetId: Get<<Self as Config>::AssetId>

Asset ID of the fungible asset “TICKET”

type InitialPercentageTickets: Get<u8>

Initial amount of Tickets that are converted as soon as FRAG are locked

type InitialPercentageNova: Get<u8>

Initial amount of NOVA that are converted as soon as FRAG are locked

type USDEquivalentAmount: Get<u128>

Amount of Tickets/NOVA equal to 1 USD